LevelCrossingDetector: convert some 64-bit temporaries to 32 because we we can't have >2^32 entries in a SSBO anyway

+++ scopehal/LevelCrossingDetector.h
@@ -97,7 +97,7 @@ protected:
 	std::unique_ptr<ComputePipeline> m_gatherPipeline;
 
 	AcceleratorBuffer<int64_t> m_temporaryResults;
-	AcceleratorBuffer<int64_t> m_gatherIndexes;
+	AcceleratorBuffer<uint32_t> m_gatherIndexes;
 	AcceleratorBuffer<int64_t> m_outbuf;
 };
